[NO-JIRA] Move status container to common
