+++ package.json
@@ -60,7 +60,7 @@
     "express": "^4.17.1",
     "express-fileupload": "^1.2.0",
     "express-slow-down": "^1.3.1",
-    "gg-components": "^12.1.0",
+    "gg-components": "^13.2.6",
     "helmet": "^3.23.3",
     "hoist-non-react-statics": "3.3.2",
     "immer": "^7.0.7",
